Revitalizing all the things! 3/5
A commit in no rush whatsoever.

+++ src/utils/Distributions.java
@@ -313,8 +313,8 @@ public enum Distributions {
 
             int[] perlinNoise = new int[currentLen];
 
-            float step = 1f / currentLen;
-            float randomStart = (float) (random.nextInt(currentLen));
+            double step = 1f / currentLen;
+            double randomStart = (float) (random.nextInt(currentLen));
             int octave = (int) (Math.log(currentLen) / Math.log(2));
 
             for (int i = 0; i < currentLen; i++) {
@@ -899,8 +899,8 @@ public enum Distributions {
 
             int[] perlinNoise = new int[currentLen];
 
-            float step = 1f / currentLen;
-            float randomStart = (float) (random.nextInt(currentLen));
+            double step = 1f / currentLen;
+            double randomStart = (float) (random.nextInt(currentLen));
             int octave = (int) (Math.log(currentLen) / Math.log(2));
 
             for (int i = 0; i < currentLen; i++) {
